refactor: i applied new structure to all files
